Move storage to its own dir
