+++ clients/functions/update-client/main.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"encoding/base64"
 	"errors"
-	"kargo-back/clients/storage"
 	"kargo-back/shared/apigateway"
 	"kargo-back/shared/normalize"
+	storage "kargo-back/storage/clients"
 	"net/url"
 	"time"
