+++ partners/functions/create-partner/main.go
@@ -3,9 +3,9 @@ package main
 import (
 	"context"
 	"encoding/base64"
-	"kargo-back/partners/storage"
 	"kargo-back/shared/apigateway"
 	models "kargo-back/shared/partners-models"
+	storage "kargo-back/storage/partners"
 	"net/url"
 	"time"
