+++ storage/partners/partner_storage.go
